rename NestedToPipeOeratorRector to SequentialAssignmentsToPipeOperatorRector
