--- rules-tests/Php85/Rector/StmtsAwareInterface/NestedToPipeOperatorRector/Fixture/skip_single_call.php.inc
@@ -1,6 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php85\Rector\StmtsAwareInterface\NestedToPipeOperatorRector\Fixture;
-
-$value = "hello world";
-$result1 = function3($value);
