+++ rules-tests/Php85/Rector/StmtsAwareInterface/SequentialAssignmentsToPipeOperatorRector/Fixture/skip_if_used_elsewhere.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php85\Rector\StmtsAwareInterface\NestedToPipeOperatorRector\Fixture;
+namespace Rector\Tests\Php85\Rector\StmtsAwareInterface\SequentialAssignmentsToPipeOperatorRector\Fixture;
 
 $value = "hello world";
 $result1 = function3($value);
