+++ rules-tests/Php85/Rector/StmtsAwareInterface/SequentialAssignmentsToPipeOperatorRector/Fixture/skip_single_call.php.inc
@@ -1,0 +1,6 @@
+<?php
+
+namespace Rector\Tests\Php85\Rector\StmtsAwareInterface\SequentialAssignmentsToPipeOperatorRector\Fixture;
+
+$value = "hello world";
+$result1 = function3($value);
